Refactor parsers to use abstract base class with template method pattern
- Add GcodeParserBase with shared functionality (RenderNoteTemplate,
  CountTemplateMatches, ParseAsSeconds, EstimateFilamentUsageInMg, etc.)
- Refactor all 5 parsers to extend base class using primary constructors
- Reduces ~1500 lines of duplicated code to ~400 lines
